requirements.txt
openai-whisper
pydub
python-dotenv
torch
tqdm
numpy